requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
Django==4.1
django-environ==0.9.0
djangorestframework==3.13.1
idna==3.3
jsonfield==3.1.0
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
pyxero==0.9.3
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
structlog==22.1.0
urllib3==1.26.12
whitenoise
intuit-oauth
python-quickbooks
